requirements.txt
flask>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
torch>=1.9.0
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.19.0
requests>=2.26.0
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.0
isort>=5.9.0 